Added junction table and the appropriate regression tests
A junction table was added to speed up the process of
searching in a graph. The junction table has only "id"
and "label_id" as columns.
In order for the junction table to have access to all the
vertices, it is created before the "_ag_label_vertex" table
and it is a parent to it.
In order to have the columns correctly inherited and in
the right order, the junction table is created with
"id", "properties" and "label_id" columns, but the
"properties" column is getting dropped after the creation
and inheritance of the "_ag_label_vertex".

Signed-off-by: Panagiotis Foliadis <pfoliadis@hotmail.com>